Move function to four_c namespace
