+++ src/meshpy/cosserat_curve/warping_along_cosserat_curve.py
@@ -35,13 +35,13 @@ import quaternion
 from numpy.typing import NDArray
 
 from meshpy.core.conf import mpy
-from meshpy.core.function_utility import create_linear_interpolation_function
 from meshpy.core.geometry_set import GeometrySet
 from meshpy.core.mesh import Mesh
 from meshpy.core.node import Node, NodeCosserat
 from meshpy.core.rotation import Rotation
 from meshpy.cosserat_curve.cosserat_curve import CosseratCurve
 from meshpy.four_c.boundary_condition import BoundaryCondition
+from meshpy.four_c.function_utility import create_linear_interpolation_function
 from meshpy.geometric_search import find_close_points
