+++ src/meshpy/four_c/dbc_monitor.py
@@ -33,12 +33,12 @@ from typing import Optional
 import numpy as np
 
 from meshpy.core.conf import mpy
-from meshpy.core.function import Function
-from meshpy.core.function_utility import (
-    create_linear_interpolation_function,
-)
 from meshpy.core.geometry_set import GeometrySet
 from meshpy.four_c.boundary_condition import BoundaryCondition
+from meshpy.four_c.function import Function
+from meshpy.four_c.function_utility import (
+    create_linear_interpolation_function,
+)
 from meshpy.four_c.input_file import InputFile
