+++ tests/test_four_c.py
@@ -32,7 +32,6 @@ input files."""
 import numpy as np
 
 from meshpy.core.conf import mpy
-from meshpy.core.function import Function
 from meshpy.core.geometry_set import GeometrySet
 from meshpy.core.material import MaterialReissner
 from meshpy.core.rotation import Rotation
@@ -40,6 +39,7 @@ from meshpy.four_c.beam_potential import BeamPotential
 from meshpy.four_c.boundary_condition import BoundaryCondition
 from meshpy.four_c.dbc_monitor import linear_time_transformation
 from meshpy.four_c.element_beam import Beam3rHerm2Line3
+from meshpy.four_c.function import Function
 from meshpy.four_c.input_file import InputFile
 from meshpy.four_c.locsys_condition import LocSysCondition
 from meshpy.four_c.solid_shell_thickness_direction import (
