+++ tests/test_meshpy.py
@@ -40,7 +40,6 @@ import vtk
 from meshpy.core.conf import mpy
 from meshpy.core.coupling import Coupling
 from meshpy.core.element_beam import Beam
-from meshpy.core.function import Function
 from meshpy.core.geometry_set import GeometryName, GeometrySet, GeometrySetNodes
 from meshpy.core.material import (
     MaterialBeam,
@@ -61,6 +60,7 @@ from meshpy.four_c.element_beam import (
     Beam3rHerm2Line3,
     Beam3rLine2Line2,
 )
+from meshpy.four_c.function import Function
 from meshpy.four_c.header_functions import (
     set_beam_to_solid_meshtying,
     set_header_static,
